+++ namedtuples.py
@@ -8,7 +8,9 @@ from typing import Optional  # for typings
 
 import tensorflow as tf  # for deep learning based ops
 
-from .losses import WasserstienLossMixin, DCGANLossMixin, LSGANLossMixin  # load loss mixins
+from .checkpoint import BaseCheckpointSaverMixin
+from .losses import (DCGANLossMixin, LSGANLossMixin,  # load loss mixins
+                     WasserstienLossMixin)
 from .utils import \
     generate_and_save_images  # for saving and generation of image
 
@@ -125,20 +127,20 @@ class BaseGANTrainer(tf.Module):
                 self.discriminator.trainable_variables))
 
 
-class WasserstienGANTrainer(WasserstienLossMixin, BaseGANTrainer):
+class WasserstienGANTrainer(WasserstienLossMixin, BaseGANTrainer, BaseCheckpointSaverMixin):
     """
     A Class for performing wasserstien gan ops
     """
-    generator_loss = True
+    pass
 
-
-class DCGANTrainer(DCGANLossMixin, BaseGANTrainer):
+class DCGANTrainer(DCGANLossMixin, BaseGANTrainer, BaseCheckpointSaverMixin):
     """
     A Class for performing DCGAN trainer ops
     """
     pass
 
-class LSGANTrainer(LSGANLossMixin, BaseGANTrainer):
+
+class LSGANTrainer(LSGANLossMixin, BaseGANTrainer, BaseCheckpointSaverMixin):
     """
     A Class for perfroming LSGAN trainer ops
     """